Fix: Vulnerability. Cache entries can be replayed.
If the clock (SystemTime) is delayed, existing cache entries will always become cache-hits. Requests will never be sent to the NetworkDispatcher if older cache entries are present. Payloads will be replayed in a dangerous transparent manner to the user and to developers who are not familiar with the bug.
Added verification for ServerDate - not foul proof, but will handle any 'accidental' SystemTime changes/mistakes on the user's side.

Signed-off-by: Henrique Marques da Silva <henriquecms@gmail.com>